Allow max_hits to be specified as "all".

+++ include/collection.h
@@ -226,7 +226,7 @@ public:
                           size_t drop_tokens_threshold = Index::DROP_TOKENS_THRESHOLD,
                           const spp::sparse_hash_set<std::string> & include_fields = spp::sparse_hash_set<std::string>(),
                           const spp::sparse_hash_set<std::string> & exclude_fields = spp::sparse_hash_set<std::string>(),
-                          size_t max_facet_values=10, size_t max_hits=500,
+                          size_t max_facet_values=10, int max_hits=500,
                           const std::string & simple_facet_query = "",
                           const size_t snippet_threshold = 30,
                           const std::string & highlight_full_fields = "",
